+++ src/def.rs
@@ -9,5 +9,5 @@ use std::io;
 
 /// Marker trait for definitions.
 pub trait Definition: Sized {
-    fn new(id: u16, buffer: &[u8]) -> io::Result<Self>;
+    fn new(id: u32, buffer: &[u8]) -> io::Result<Self>;
 }
