+++ src/ldr/rs3.rs
@@ -7,7 +7,7 @@ use crate::def::rs3::ItemDefinition;
 /// Loads all item definitions from the current cache.
 #[derive(Clone, Eq, PartialEq, Debug, Default)]
 pub struct ItemLoader {
-    pub items: HashMap<u16, ItemDefinition>
+    pub items: HashMap<u32, ItemDefinition>
 }
 
 impl_rs3_loader!(ItemLoader, ItemDefinition, items, archive_id: 19);
